+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+##  - 2022-05-25
+
 ## 2.2.1 - 2022-04-04
 
 ### Fixed