+++ CHANGELOG.md
@@ -5,9 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- add parameter `--all-topics` to allow reset consumergroup for all topics. See [#127](https://github.com/deviceinsight/kafkactl/issues/127)
 
 ##  - 2022-05-25
-- add parameter `--all-topics` to allow reset consumergroup for all topics. See [#127](https://github.com/deviceinsight/kafkactl/issues/127)
 
 ## 2.2.1 - 2022-04-04